Readd vgit. Toggle variable for maxing windows.
Readd vgit for the coloring in the gutter to show what lines were
modified, as well as to automatically show the commit message for the
current line.

Switch maximzing a window to use a global bool, in order to be more
reliable. Maximzie / unmaximze are available to be called as functions.